.github/scripts/requirements.txt
PyGithub
groq


